relocate timeslot files
